BOLT 12: the signature on offers can be omitted.
It was pointed out that often low-end camera resolutions can be
limited, so I thought hard about how to shrink offers.

This is by far the largest field, and does not seem required (the
invoice, of course, is signed).  It weakens the case to omit the
bech32m checksum though.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>